requirements.txt
astroid==2.2.5
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
numpy==1.17.0
pandas==0.25.0
Pillow==6.1.0
pylint==2.3.1
pytesseract==0.2.7
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
yapf==0.28.0
